phishguard_ai/requirements.txt
django>=4.2,<5.0
requests>=2.31.0
python-dotenv>=1.0.1
gunicorn>=21.2.0
whitenoise>=6.6.0
urllib3>=2.2.0
